requirements.txt
requests>=2.28.0
gitpython>=3.1.30
python-jenkins>=1.7.0
pyyaml>=6.0
jinja2>=3.1.2
argparse>=1.4.0
python-dateutil>=2.8.2
urllib3>=1.26.0
cryptography>=39.0.0
python-dotenv>=1.0.0 